gps alt now float m

+++ Zephyrus/FC/FC.ino
@@ -360,7 +360,7 @@ void constructTelemetryPacket() {
   uint8_t gpsFix = gps.getFixType();
   int32_t lat = gps.getLat();
   int32_t lon = gps.getLon();
-  uint32_t alt = gps.getHeight();
+  float alt = gps.getHeight();
   uint32_t hACC = gps.getHAcc();
   uint32_t vACC = gps.getVAcc();
   uint8_t numSat = gps.getNumSV();
